Updated firmware for SB2:
Fixed bug 958 where  when the link goes down (wireless or ethernet ) we don't automatically reconnect.  
In part this was due to not checking tcp_connection_connect result code.   Now we do.
Now ethernet should connect within a second.  Wireless may take ten extra seconds, since the result
code from tcp_connection_connect on a down wireless connection doesn't indicate any problem, so we set up our 
own timer to time out establishment.

Fixed a bug where the first character entered in an ssid or wap key would be wrong.

Save the player's IP address as its name.  The hello packet should (in the future) provide
a human readable name for the player to cache.

When I disabled squeezenetwork menus, i added an off-by-one bug where the last server in
the list would appear as squeezenetwork.  should be fixed.

Fixed crashing bug when switching from wired to wireless.   We need to reboot when the user 
does this so we can turn on the wireless interface.  Crashing was due to scanning without a network card.

Fixed a problem where the reconnect bit could be wrong when switching between servers and/or the SqueezeNetwork.
Now when we switch servers, we reset the slimproto_have_connected bit.




git-svn-id: file:///home/awy/mirror/slim/trunk/server@2635 62299810-d8cb-41fd-93b7-d32162e5a4a4