Look at the 'Date Added' from iTunes and the mtime of the file on disk to determine if we even need to continue updating the track. This is the big win - we don't even fetch/update from the database if we don't need to.
Fix the case where this is an initial scan with no pref file. Don't start
scanning when initPlugin() is called, which will call doneScanning(), and
reset some of the values we need.

Add a markAsInvalid to DBIStore, also updating the zombieList to use URLs.

Turn off the DB garbage collection for now - lazily delete tracks from the
DB. I still need to turn back on the Album / Contributor / Genre sweep
though, but in such a way that it doesn't happen every time iTunes is updated.

There's also a fix for bug #975, where URLs (and iTunes playlists) were
creating the "No Album", "No Artist" DB entries.



git-svn-id: file:///home/awy/mirror/slim/trunk/server@2642 62299810-d8cb-41fd-93b7-d32162e5a4a4