New firmware revision including: 1) Fixed leak in setup that caused us to crash eventually if we were disconnected from a server. 2) New reconnect behavior described below.
Improved reconnect behavior:
1) Changed the definition of the reconnect bit in the HELO message for
SB2 to mean there is an active data connection.
2) If there is no active data connection, SB2 sends the last
bytes_received value. This can be used by the server to reconnect to
the box seamlessly. (The protocol may change to be a lot cleaner in
the future, since the HELO message may not be the cleanest place for
it).
3) Changed the loop_song field of the control header to be a bit flag. 
There is a still a looping bit. There's also a bit to indicate that the
connection should not restart the decoder. This bit is now used by
the server to reconnect seamlessly to the SB2 by seeking to the point
before the drop in connection and restarting from there.



git-svn-id: file:///home/awy/mirror/slim/trunk/server@2659 62299810-d8cb-41fd-93b7-d32162e5a4a4