Better handling of tracks & zombies:
Don't limit the file types that objectForUrl when passing onto newTrack can
handle. newTrack is now updateOrCreate as well, because it's possible that a
url can be put on the zombie list, and then used again before the garbage
collector has gotten to it.

updateOrCreate now unconditionally removes items from the zombie list as well.

In getCurrentTitle - if we have an undef url, don't even try to fetch or go to
standardTitle() - this should alleviate bug 1158



git-svn-id: file:///home/awy/mirror/slim/trunk/server@2683 62299810-d8cb-41fd-93b7-d32162e5a4a4