Fix bug1124: now playing (jump back on wake) not update properly
if a mode is currently animating, then the server has no way of
entering a screensaver.  This was by design, since animation was a
screensaving tactic. However, with preferred modes for screensavers,
it makes more sense to go into the preferred mode regardless of other
animation.

Blocking still takes precedence, and we will start a scrollBottom if
the current mode is not currently animating.



git-svn-id: file:///home/awy/mirror/slim/trunk/server@2686 62299810-d8cb-41fd-93b7-d32162e5a4a4