fix bug1168: scanned playlists block scan from an empty db
it seems that the checker for musicmagic was triggering a scan, or
delaying long enough to allow a player connect to trigger playlist
load.  This beats the datasource check from slimserver.pl, so a scan
wont start if there is an empty database. 

delaying the checker solves this.  Its not a bad idea to put it off
from startup anyway, so do this to MoodLogic as well.




git-svn-id: file:///home/awy/mirror/slim/trunk/server@2698 62299810-d8cb-41fd-93b7-d32162e5a4a4