Update the importer code:
addImporters() now takes a hash ref for named params.

Additional function for 'reset' - the only caller right now is iTunes, which
for a wipeDB, we want to reset the lastITunesMusicLibraryDate pref, so the
entire XML file is read - not just what's been changed.

Rescan Plugin should tell the importers that it's a rescan so the DB cleanup job can be run.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@2734 62299810-d8cb-41fd-93b7-d32162e5a4a4