Fix for iTunes playlists with apostrophes in their names coming up empty. We weren't consistently escaping the URLs.
Bug: 1226
Description:

The 'lastITunesMusicLibraryDate' pref served two purposes - it was
used to determine when the iTunes XML file had changed, but it was
also an indicator that we'd already done an iTunes import (and
shouldn't bother about files whose modification time were older than
the XML file). The problem is that many people had an old value of the
pref from prior to SlimServer 6.0. The result was that we sometimes
didn't scan the entire iTunes library.

The fix is to use a new pref name for this purpose. We also
reinitialize the pref on a Rescan (we were previously only doing it
for a Wipe Cache) so that a manual rescan is more thorough.



Bug: 1278
git-svn-id: file:///home/awy/mirror/slim/trunk/server@3009 62299810-d8cb-41fd-93b7-d32162e5a4a4