We weren't taking TSOP-based sort information if the artist object had already been created without a TSOP tag. The new behavior is last TSOP wins.
Bug: 1306
Description:
We were inconsistently keying on the artist name and namesort depending
on whether the artist came out of the in-memory cache or the database.
This resulted in duplicate artist entries in some situations.

Bug: 1382
Description:
$client->currentPlaylist() not updated with synced players, resulting
in incorrect display behavior.

Also, added a fix that makes sure that the initial song queue for players
only has one entry.



Bug: 1005
git-svn-id: file:///home/awy/mirror/slim/trunk/server@3058 62299810-d8cb-41fd-93b7-d32162e5a4a4