Bug: 1241
Description: when regenerating the playlist html objects,
$client->param('start') is defaulted to zero.  pageBar, however uses
undef to know when a url form the web UI has triggered the playlist
refresh.  Instead, use the same bit of code that pagebar uses to set
the default.  This way, the page reload will start on teh right page
on refresh, and when turning shuffle off.



Bug: 1483
git-svn-id: file:///home/awy/mirror/slim/trunk/server@3153 62299810-d8cb-41fd-93b7-d32162e5a4a4