Make more widespread use of Pages.pm $fieldInfo. 	      This now allows *_mix_list.html from the imports plugins 	      to be dumped in favour of using browsedb_list.html with 	      [% webroot %] properly used.
	      Livesearch is still an entity unto its own due to heavy
	      hardcoding.

Bug: 1446

Description: Musicmagic shoudln't scan when it is still analysing
music, so Predixis added a new getStatus API that the server can use
to defer a scan when the library is changed.  Also move last scan time
to a server pref, and also make use of the new contents param for the
cacheid API that will no longer cause a rescan on startup due to state
changes in musicmagic that do not matter to slimserver.

Will need to be tested by those having the troubles to be sure how
much this helps relieve the cpu issues




Bug: 1180
git-svn-id: file:///home/awy/mirror/slim/trunk/server@3272 62299810-d8cb-41fd-93b7-d32162e5a4a4