SlimServer.exe is dead. Long live SlimTray.exe.
Killed SlimServer.exe, the Windows GUI application that has recently
been causing much pain to our tech support crew. Its new incarnation
is a tray tool that can be used to check the status of SlimServer,
stop and start the service, and launch the web interface in the user's
default browser. SlimServer is now always installed as a Windows
service (the option of starting automatically at boot time still
exists).



git-svn-id: file:///home/awy/mirror/slim/trunk/server@3480 62299810-d8cb-41fd-93b7-d32162e5a4a4