Moved to firmware 14 to deal with the following direct streaming problem:
We weren't consistently sending the empty body message at the end of a
direct streaming playlist body. The result was a server that wouldn't
start playback of a radio stream. This change fixes a couple of
codepaths that could trigger this condition. Still need to talk to
Dean about the direct streaming logic - specifically when there isn't
a Content-Length header in the response header for the playlist.

Fixed direct streaming on server to allow for multiple empty body
messages.



git-svn-id: file:///home/awy/mirror/slim/trunk/server@3565 62299810-d8cb-41fd-93b7-d32162e5a4a4