Merged in r3771 from 6.1.x branch:
ix for bug introduced when we started returning LightWeightTrack objects
from objectForURL if one was passed in (r3740). This prevents people from
transcoding to MP3 or doing a FFWD/REW.

The problem is that someone made a bitrate() method for Track that takes
different arguments than the default bitrate accessor. LightWeightTrack
also has an accessor, so the inherited method isn't invoked. The fix is
to explicitly invoke the super class method.



git-svn-id: file:///home/awy/mirror/slim/trunk/server@3780 62299810-d8cb-41fd-93b7-d32162e5a4a4