infoFormat is dead! Long live infoFormat!
Replaces incomprehensible regex with something hopefully a bit easier to work with.
Old stuff left in place, but renamed to *_old for head to head testing.  Will clean up later.

Basically, the title format is parsed once and turned into a function which takes the track object as a parameter, and returns the formatted string.
These functions are cached in the %parsedFormats hash for reuse.



git-svn-id: file:///home/awy/mirror/slim/trunk/server@3784 62299810-d8cb-41fd-93b7-d32162e5a4a4