Updates to how we parse the iTunes xml file and playlists on Windows.
I was able to reproduce a case which would generate duplicates and bad
database entries that had a > 127 character in the path.

iTunes stores the files as url encoded UTF-8. Which we handle properly. But we
add the track to the playlist table too soon - so bogus entries are generated.

A similar thing was happening with playlist files - two problems there:

Playlist created in Winamp pointed to:

c:/temp/music/foo.mp3

The real path, and the one we build from opendir(), etc is:

C:/Temp/music/foo.mp3

So do proper normalization of the path & volume name in fixPath()

Also - we don't want to encode incoming playlists that are latin1 - otherwise
they are encoded incorrectly, and cause duplicate entries.

Bug: 1786
git-svn-id: file:///home/awy/mirror/slim/trunk/server@3801 62299810-d8cb-41fd-93b7-d32162e5a4a4