Fixes for playlist / cue sheet parsing.
The changes to seek() from sysseek() the other day also require a change to
use tell() for finding the position. This was causing a test cue sheet to fail.

Print out the useful url instead of a FileHandle GLOB for playlist format debugging.

Don't try and create / readTags for a cuesheet without a FILE.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@3818 62299810-d8cb-41fd-93b7-d32162e5a4a4