Squeezebox2 Firmware 16:
Remove reset of MAC address when (accidentally or intentionally) entering factory test mode.

Bug 1200: Add WOL feature.  Sends a WOL packet to SlimServer on local subnet when connecting 
via the top-level setup menu or by pressing POWER or RIGHT when in the "Connecting to server..." screen.

This should be moved to the 6.2.x branch once it's been tested a bit.



git-svn-id: file:///home/awy/mirror/slim/trunk/server@3854 62299810-d8cb-41fd-93b7-d32162e5a4a4