Simplify Live_Object handling in Class::DBI
Integrate some code from Class::DBI 0.99 to make the live_object storing and
fetching much more straight forward.  

Catch when we have a ::Deleted object, and refetch it.

Add a _class_for_ref() method which allows us to not have both a ::Track and
::LightWeightTrack in the object index.

In ::Pages - we might need to inflate an object for a client playlist.



git-svn-id: file:///home/awy/mirror/slim/trunk/server@3904 62299810-d8cb-41fd-93b7-d32162e5a4a4