Fix searching when one has a specified sort order tag.
Use the (name|title)search columns that were added for this purpose in change 3849

Store a canonical version of the regular (name|title) in this column.

Update all callers to use the column for matching.

Remove confusing cache code in ContributorTrack & GenreTrack, now that the
Live_Objects code in Class::DBI is working properly.


Bug: 1636
git-svn-id: file:///home/awy/mirror/slim/trunk/server@3938 62299810-d8cb-41fd-93b7-d32162e5a4a4