Remove no-longer used elemLookup() and addToInfoHash() functions. They were left-overs from the old infoFormat function.
Add addFormat() function.  This allows new elements or formats to be added to the %parsedFormats hash.

This allows for custom elements or optimised formatting to be added by plugins.  It takes as parameters the format string or element, and a reference to a subroutine to replace that string or element.

The referenced subroutine will be passed a track object as a parameter and should return the formatted string.




git-svn-id: file:///home/awy/mirror/slim/trunk/server@4100 62299810-d8cb-41fd-93b7-d32162e5a4a4