Changes to Common Album Titles:
The current method for listing album titles isn't scalable, and was put in
place for the edge case of a Multiple Artist album with a Common Title.

Add a checkbox which defaults to off, which will work for 99% of the people out there.

We'll match albums (on directory boundaries) with both contributor & album title.

Additionally, only add "tracks" that are audio to the lastTrack hash, which
will provide better matching heuristics.

Bug: 1378, 2023
git-svn-id: file:///home/awy/mirror/slim/trunk/server@4111 62299810-d8cb-41fd-93b7-d32162e5a4a4