From TheEndless:
After going through the "XML Skin" tech info documentation, I
realized that it explicitly states that the player id has to be added to
URLs targetted at specific players.  I had changed the skin to
automatically add the player id, but decided to change it back to avoid
breaking any existing applications that may have been appending the
player id based on the orignal rules.  Having the application add the
player id also allows it to make a single XML query while targeting
multiple players.  The only exception are playlist specific URLs (in
status.xml) which wouldn't really apply to anything but the specified
player.

I also added genre_id, artist_id, and album_id elements to the song
info XML.  Ideally the info for genre, artist, and album would be
implemented as a single elements with id and url attributes, but for
simplicity and backwards compatibility, I just added a new element for
each.



git-svn-id: file:///home/awy/mirror/slim/trunk/server@4419 62299810-d8cb-41fd-93b7-d32162e5a4a4