create a central template for common hrefs and links in EN.  This is now called hreftemplate and can be PROCESSED as part of any skin file that wishes to make use of it.  This helps to remove any hrefs and web commands from skins and centralize things that could potentially change as the server changes command structure in the future. More to be added to hreftemplate as needed.
Fishbone and Default mostly converted.  More commands can be added to
hreftemplate as needed.  

This can lead to skins providing warppers for
these commands so that plugin templates need only to provide a single
template that provides command params and leaves the skins to do
layout and the EN template for the final commands.  



git-svn-id: file:///home/awy/mirror/slim/trunk/server@4430 62299810-d8cb-41fd-93b7-d32162e5a4a4