Snow updates from James Craig
"Thanks for the tips - I had deleted U in error when playing about 
with the characters! I've also narrowed the V a little - it was so 
much wider than the other characters. To change the messages, edit 
the SNOW_WORD strings. And if you change the number that exist for 
a given language you would need to change the NUMBER_OF_WORDS as 
well. Attached should be my latest version which also fixes reversed 
flake positions on the SBG, and (again) increases the amount of snow."



git-svn-id: file:///home/awy/mirror/slim/trunk/server@4434 62299810-d8cb-41fd-93b7-d32162e5a4a4