fix to checkDefaults where same pref was being written for both.     - slightly altered patch from max.  bt() moved inside plugins     debug. notes from max:
"The attached patch adds genre filtering to random mix as well as making
the ui conform to the new user interface guidelines.  I've added a sub
menu where the user can choose from all the genres in their library. 
The mixes then limit themselves to the genres they've selected.  The
choices are stored in a client pref so are persistent."



Bug: 2229
git-svn-id: file:///home/awy/mirror/slim/trunk/server@4508 62299810-d8cb-41fd-93b7-d32162e5a4a4