updates from max:
Here's a patch that adds the show briefly behaviour that I was talking
about in the last post.  It makes the previous patch obsolete.  This
does the following:
- Removes the "press play to stop" behaviour on the menu
- When mix is playing, changes that mixes menu item from Random Song
Mix to Playing Random Songs etc
- Removes the extra update when pushing into the plugin
- Catches empty plugin_random_number_of_tracks pref and sets it to 10
(this cleans up the effects of a previous bug)
- Show briefly now happens when pressing add (don't know what fixed
this!)
- Adds show brieflys to say what's being added as it happens e.g.
Adding to end of playlist/Random Album: Are You Experienced?



git-svn-id: file:///home/awy/mirror/slim/trunk/server@4536 62299810-d8cb-41fd-93b7-d32162e5a4a4