Stop using the Class::DBI Live Object Index, which is a hack.
Somehow the album member of a track was being set to {}, which is an unblessed hash ref.

I haven't been able to track down where that's being set, but turning off the
Live Object Index (and Is_Weaken_Available specifically) fixed the issue.

Also, check for a busted value in the HasA code that was actually crashing as well.

I can't wait to get rid of Class::DBI.


Bug: 2368
git-svn-id: file:///home/awy/mirror/slim/trunk/server@4986 62299810-d8cb-41fd-93b7-d32162e5a4a4