bug 2494: start extracting command code into small functions, and rewriting execute() to look up those functions in a table.
I used 'pref' as a simple proof-of-concept, and as far as I can tell, it
works: it sets the preference properly, and returns the same array as the
old code. however, the new code is disabled by default. for migration, I'd
imagine replacing the if..else cases in execute() one by one with calls to
execute_new(), as the commands get refactored into functions; then at some
point, cut over completely.



Bug: 2494
git-svn-id: file:///home/awy/mirror/slim/trunk/server@5050 62299810-d8cb-41fd-93b7-d32162e5a4a4