Now that we've turned off the "Live Object Index" cache in Class::DBI, we are no longer guaranteed to get the same object in memory when doing an objectForUrl (or objectForId). Thus, we should pass in objects to functions that might need to modify them.
Through the twisty-turny passages that are the Slim::Music::Info functions, if
we pass along a track object to updateOrCreate when updating the Content-Type
for a URL, we'll use the existing object.

This is really all a big mess and is still left over from the pre 6.x days.

We need a generic framework to make HTTP requests (use AsyncHTTP?) which
doesn't rely on having anything in the database - so that callers such as the
MusicMagic plugin can use it to check if the MMM API port is open. 

Then on top of that, have a caller that will retrieve data from HTTP (headers
included), and manipulate it at the database/track level if required.

Said modules should also not be located under the Slim/Player/Protocols/ hierarchy.


Bug: 1811
git-svn-id: file:///home/awy/mirror/slim/trunk/server@5079 62299810-d8cb-41fd-93b7-d32162e5a4a4