add setting in web and player UI for a 20s fade-in for the alarms.  This is a per-player toggle acting globally for all built-in alarms for the given player.
clean up a slipped whitespace from a fix done last night in BrowseDB

for the alarm pref checkbox, create a new sub in Common called
checkBoxOverlay.  It takes a single boolean as an arg, which may be a
server pref.  This way, when Common.pm becomes a class, we can have a
built up set of common UI widgets with this one included.



Bug: 2575
git-svn-id: file:///home/awy/mirror/slim/trunk/server@5222 62299810-d8cb-41fd-93b7-d32162e5a4a4