Big update that does 2 major things:
* In AsyncHTTP, both DNS lookups and connect() have been made non-blocking.  Most HTTP code will benefit greatly from this, including XMLBrowser, 
Live365, Shoutcast, and Pandora on SN.  Some HTTP code that doesn't use AsyncHTTP such as RssNews will still block.  That will hopefully be fixed 
in a later update.

* In Scan, added a checkConnect method that uses AsyncHTTP to check connection to remote stream servers before they are added to the playlist.  
This prevents blocking from happening in RemoteStream's connection code, which is not yet non-blocking.  This also adds an error message 
when a stream doesn't connect.



git-svn-id: file:///home/awy/mirror/slim/trunk/server@6706 62299810-d8cb-41fd-93b7-d32162e5a4a4