Patch from Chris Key to alter number key navigation in numerically sorted lists:
"This patch modifies the behaviour when a user selects items in a list with the digit keys.  For 10 items or less, the functionality is unchanged, i.e. 1-9 maps to 1-9 and 0 maps to 10.  For 11 or more items, the user can enter a number directly, e.g. the input sequence 1,4,2 would select indexes 1,14,142 in turn.  The timeout before new key presses start a new number is the same as when using numbers to enter letters.  If a user enters a number beyond the end of the list, they are taken to the last item in the list.  To maintain similarity to the scheme for 10 or less items, if a number is entered starting with 0, a preceding 1 is implicit, e.g. the sequence 0,2 would yield index 102"



git-svn-id: file:///home/awy/mirror/slim/trunk/server@6868 62299810-d8cb-41fd-93b7-d32162e5a4a4