Modify array handling in setup a bit.
Added preprocessArray function which is always called for array prefs.  This figures out arrayMax, and trims off any extra, unused array options.
	This removes the need to call removeExtraArrayEntries from a preEval section.

Modified processChanges to always call processArrayChange for array prefs which do not have 'dontSet' set.  This, and the 'onChange' for the array, are only called once.
	This removes the need to have an 'onChange' section where all you do is call processArrayChange.  It also removes the need to check the changed hash for a 'Processed' key.

Changed RssNews and Podcast plugins to make use of the new array handling.

Both of these should be backwards compatible with the existing way of doing business.

git-svn-id: file:///home/awy/mirror/slim/trunk/server@6941 62299810-d8cb-41fd-93b7-d32162e5a4a4