Really discard key presses if they have been queued up for too long to avoid UI queuing up actions when server is blocked.
r6283 attempted to do this, but it turns out that most of the queueing
actually occurs in the slimproto tcp session when the server does not
visit select for a long time.  This code estimates the time the IR code
is generated and hence is able to detect keypresses stored in the slimproto
socket for a long time.

IR queue moved to single queue of hashes to ease maintainance


Bug: 2081
git-svn-id: file:///home/awy/mirror/slim/trunk/server@7456 62299810-d8cb-41fd-93b7-d32162e5a4a4