Merge from 6.3 (7940:7942)
Fix a race condition in quickstart where it could unpause too early if the buffer fullness hadn't been reset by a recent STAT call.
Call quickstart() more frequently.
Changed the logic in stream() when command is 's', so we don't send a request string for other types of commands such as unpause.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@7943 62299810-d8cb-41fd-93b7-d32162e5a4a4