Avoid performance problems with newer versions of perl5.8.8 - as seen with FC5 perl5.8.8-5.
Recent versions of perl 5.8.8 include a bug fix for which makes
reblessing significantly slower.  The DBIx/Class code was triggering
this due to the way the rs object was blessed.  Blessing a named hash
seems to avoid this.

Not moved to lib as Dan plans to update to newer version of DBIx/Class soon.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@8581 62299810-d8cb-41fd-93b7-d32162e5a4a4