Bug: 2647 (partial) Description: rework of settings UI.  This makes much greater use of the checkBox UI and fits the settings into the home menu API.  This means that modules can programmatically add settings items by using Home::addSubMenu.  See AlarmClock and Information for examples.  The 'condition' param allows the module to set conditions on which the setting will appear (eg Transporter only, or when ReplayGain is available).  The TextSize pref still needs to be converted but this will take a bit more work in regard to font naming.
Some updates to INPUT.Choice to handle 'pref' items, making the checkbox as automatic as possible.  Users of INTPUT.Choice should be able to build with the UI and remain consistent should it ever change.

screensavers are now REQUIRED to provide a string token only as the displayname param in addSaver.  Most appear to be ok, but MusicInfo and XNowPlaying are still using pre-localised strings and will no appear correct in the player UI from this point on.



Bug: 2647
git-svn-id: file:///home/awy/mirror/slim/trunk/server@8906 62299810-d8cb-41fd-93b7-d32162e5a4a4