Changes to HTTP.pm to provide better support for non HTML/template based interfaces.
We can now have rawhandlers, much like web page handler today, but that are called very early in the
request handling code, before we've managed cookies, Icy stuff, parsed params, etc.
As a result I had to re-order the request processing a bit in processHTTP. It should not have
any adverse effect. In any case, I left comments describing the old vs new order at the start of the
procedure. It is much easier to read it there than figuring it out from diff...

The other change is the ability to use Transfer-Encoding:chunked when sending responses. This is mainly
for raw handlers. The current page handler still send Content-Length:xxx HTTP 1.0 responses.

Finally, there now a way to have a callback on connection closing. This is a generic mechanism for
raw handlers in case they do manage some internal data related to a connection.

The typical use case for raw handlers is the (now defunct) XML RPC plugin.



git-svn-id: file:///home/awy/mirror/slim/trunk/server@11502 62299810-d8cb-41fd-93b7-d32162e5a4a4