SlimServer <-> SqueezeNetwork pref syncing
When a new player connects, it will request a "sync down" of prefs from SN.
All pref changes are timestamped on both sides, so any prefs changed on SN after
the pref was changed on SS will be synced to SS.

Any time a pref is changed that SN cares about, a 30-sec timer will be set to "sync up"
that pref to SN.  This will allow rapid pref changes (i.e. volume) to get bundled up into
one operation.

None of this is enabled if SN login info is not present under Server Settings -> SqueezeNetwork
and the pref to enable syncing is set.


git-svn-id: file:///home/awy/mirror/slim/trunk/server@12348 62299810-d8cb-41fd-93b7-d32162e5a4a4