Facility for allowing custom deletion of menu items via plugins to the Jive Main Menu, or in submenus Settings, My Music, and Extras . The menus available for customization will probably be expanded going forward (e.g., Search).
It is also possible to delete a menu as provided by Slim::Control::Jive and then develop a 
customized replacement for that item to be inserted in its place, all from a plugin.

Note: if there is a menu item that is also created by Jive through an applet (e.g., Extras, Settings),
it is NOT possible to delete the menu item through this mechanism. Doing so will remove anything
delivered from SC to this menu, but not the menu item itself (or anything delivered from Jive to that
menu)



git-svn-id: file:///home/awy/mirror/slim/trunk/server@14076 62299810-d8cb-41fd-93b7-d32162e5a4a4