Do not allow player to be named '0'
this is a workaround for a jive.SlimBrowser issue with text entry input, which I've discovered
these types of input sends the cli command  twice, first when entering the text input window (with a blank entry)
and the second when inputting the text. If you exit the text entry window with the back key, you
are left with whatever the bogus cli command you issued when entering the text input window reaped.



Bug: 6384
git-svn-id: file:///home/awy/mirror/slim/trunk/server@16470 62299810-d8cb-41fd-93b7-d32162e5a4a4