New Transporter firmware, plus some minor SlimServer tweaks to support the new features.
V.40 Firmware changelog:

- New programmable logic: Xilinx CPLD will be automatically upgraded,
but if you subsequently downgrade to an earlier firmware, it must be
manually re-flashed to the older rev (press '1' on startup).

#4712
- Add support for native 88.2 playback

- Add "effects loop" feature for hooking in an external DSP via SPDIF

#4463
#4562
- CPU optimizations in UART and IR blaster, freed up over 20 MIPS.
Should prevent 24/96FLAC from ever being able to max out the CPU, even
with replaygain and spectrum analyzer enabled.

- In case the CPU does get maxed out, which it won't, throttle the
visualizer frame rate so as to give more cycles to the decoder

#4322
#4436
#4634
#4834
#4895
- Fix many bugs concerning digital input switching, white noise, and
loss of sync. There were a whole bunch of intersecting issues here.

#4806
- Could not reproduce this bug (pops when using word clock). This is
either fixed now, or was not a valid bug in the first place.

- Word clock and effects loop settings take effect immediately - no need
to reconnect or restart the track.

- Increased test tone resolution from 16 to 24 bit

- Add test tones for 88.2 and word clock

#1397
- Halt all clocks when disabling s/pdif, instead of just masking off the
s/pdif data. This results in a true 0 VDC level and should address the
remaining receivers that thinnk we're still active.

- Changed "fixed digital levels" to fix both digital and analog levels.
This was necessitated by an architectural change to support 88.2KHz.

- Due to the aforementioned change, it is no longer possible to disable
s/pdif while allowing the analog outputs to continue running. However,
the only place where we were doing this was in the event of a Rhapsody
DRM track while in aes/ebu mode. Now we just fall back to s/pdif
instead, which supports the copyright bit.

- Make word clock and effects loop settings take effect immediately
instead of on Client's initial connection to SlimServer.



git-svn-id: file:///home/awy/mirror/slim/trunk/server@17059 62299810-d8cb-41fd-93b7-d32162e5a4a4