Full LineIn plugin support in controllerUI
Line In item appears in home menu when something is plugged into the line in jack
Line In item is removed from home menu when something is removed from the line in jack

Settings->Audio->Line In Level added. Uses a slider widget to set volume and a showBriefly for UI feedback on the level that's been set
Settings->Audio->Line In Always On added. This is delivered as a "choice" item


Bug: 8318
git-svn-id: file:///home/awy/mirror/slim/7.2/private-branches/boom/server@22598 62299810-d8cb-41fd-93b7-d32162e5a4a4