add 'timeout' preference to MusicIP, to be used in synchronous requests like the mix command.
Please note: as mentioned before this can lead to hiccups if you set this value too high and other players starve waiting for MIP to return. This preference isn't exposed in any UI, as using it might negatively influence SC behaviour. It's for you poweruser's (and my ;-)) convenience only. 

To set the parameter either:
- shut down SC
- add "timeout: 10" (or whatever seconds you want) to prefs/plugin/musicip.prefs
- restart SC

or :
- "telnet yourmachine 9090"
- type "pref plugin.musicip:timeout 10" (or whatever), hit enter

Up to you to find a value which fits you well.

BTW: I'm running MIP against a collection of about 15k tracks.
- main Linux server (1GHz Via C3) took more than a week(!) to analyze this. Mixing never was an issue with the default 5 seconds.
- test Windows Home Server (1.6GHz Core2 solo - what a nice name for a good ol'Celeron!) finished the analysis in a few hours, but times out on many mixes...

Bug: 10760
git-svn-id: file:///home/awy/mirror/slim/7.4/trunk/server@28422 62299810-d8cb-41fd-93b7-d32162e5a4a4