fork any request that comes in with xmlBrowseInterimCM param to a special three item CM that only has controls for Add to End/Play Next/Play
note: the Play Next in album context (e.g. Rhapsody or Napster album) does not work yet. This will have a separate bug filed against it.

XMLBrowse CMs should be richer than this in content, but this is likely going to be where it stops for initial 7.4 release.

This fix only deals with playable items in the XMLBrowse world. Rhapsody artists, for example, do not have CM support as they are not playable items.


Bug: 14061, 13969
git-svn-id: file:///home/awy/mirror/slim/7.4/trunk/server@28573 62299810-d8cb-41fd-93b7-d32162e5a4a4