bug 10889: Place configurable upper limit on playlist length Add maxPlaylistLength server preference with default of 500 (100 on TinySC) and WebUI under Advanced / Performance. Consolidate all manipulation of player playlists into Slim::Player::Playlist. New method Slim::Player::Playlist::addTracks() used for adding tracks to playlist. Slim::Player::Playlist::removeTrack() extended to be able to remove multiple consecutive tracks.
If the playlist is full when attempting to add tracks, then tracks in the (possibly-shuffled) playlist
before the 'current' (possibly playing) track will be removed to make room. After that, if the
playlist is still full and the attempt is an insert, rather than an append, then tracks will be
trimmed from the end of the playlist as far as possible.
An error message (show-briefly) is shown if not all tracks are added.

Bug: 10889
git-svn-id: file:///home/awy/mirror/slim/7.5/trunk/server@28991 62299810-d8cb-41fd-93b7-d32162e5a4a4