instead of forking the xmlBrowseInterimCM requests off to a subroutine that delivers only 3 or 4 items, instead deliver these items prepended to the rest of the XMLBrowse response. The result is that Add Next/Add to End/Play and Favorites are all available in the CM, plus all additional information that can be delivered on that item.
Tested against internet radio items, Pandora, Rhapsody, Napster, Slacker, and Last.fm. All work as expected.

there is an underlying bug that if the CM items > BLOCK_SIZE the chunked response will result in a broken CM menu on the squeezeplay side because the count will be off. The fix for that is much more involved, and since BLOCK_SIZE is 200 and we're dealing with CMs only here, it's highly unlikely we're ever going to hit this condition.


Bug: 14145
git-svn-id: file:///home/awy/mirror/slim/7.4/trunk/server@29945 62299810-d8cb-41fd-93b7-d32162e5a4a4