Fixed bug 15150: My Apps not supported via SbS should switch player to mySB
Add 'SN' plugin to auto-switch players to SN (mySB) when attempting to play a track that cannot be played by the current server but which can be played by SN. The switch only happens when attempting to play the track, not when it is added to the playlist. The full sync-group will be moved if necessary. As much as possible of the current playlist is copied to SN, not just the one track that initiated the switch. Odd things are likely to happen with mixed local/remote, shuffled playlists.

Install the plugin in TinySC and not on SN itself.
No longer install RhapsodyDirect plugin in TinySC.

Currently has fixed list of SN-supported protocols; needs to get this dynamically.

Might be nice to get track metadata for tracks being hosted by this plugin while they are in the SbS playlist.

Bug: 15150
git-svn-id: file:///home/awy/mirror/slim/7.5/trunk/server@29991 62299810-d8cb-41fd-93b7-d32162e5a4a4