bug 16006: Ogg streams metadata not updated on SqueezePlay UIs Modify Slim::Music::Info::setCurrentTitle() to send out all the various notifications about a new title (if passed a Client parameter), effectively moving this functionality from Slim::Music::Info::setDelayedTitle().
In Slim::Player::Protocols::HTTP::parseMetadata():
Include the code for handling a stream artwork-URL along with that for a new
stream title, and set them both in a delayed callback as appropriate.
For Ogg metadata, also call setCurrentTitle() in the delayed callback.
Don't strip out CRLF from metadata before deciding what type it is, as this can
mess with the Ogg length-value formatting.
Modify the CRLF stripping to replace any sequence of whitespace including CR
or LF with '; '.

In Slim::Player::Protocols::MMS::parseMetadata(), also call setCurrentTitle()
in the delayed callback.

Bug: 16006
git-svn-id: file:///home/awy/mirror/slim/7.5/trunk/server@30519 62299810-d8cb-41fd-93b7-d32162e5a4a4