Fixed bug 3161: Need more aggressive auto-retry for internet radio
If a radio station (non-finite stream) succeeds in playing for at least 10s and then subsequently fails then, after the immediate retry (existing situation), introduce a retry mechanism. This will retry the connection at successive intervals of 5s, 10s, 15s, 30s, 30s, ... until an overall timelimit is reached. This limit is 5 minutes if the station is the only (or last) item in the current playlist and 30s if there are more tracks in the current playlist (including via repeat-playlist if that is enabled), in which case playback will advance to the next track when the 30s timeout expires.

The behaviour is slightly different between ip3k and SqueezePlay players, as ip3k players wait for the 10s connection timeout even when the TCP connection attempt is immediately rejected.

Note: this will not attempt to retry a connection if the initial connection fails or if it does not succeed in playing for 10s. It will not retry pseudo-radio services such as as Pandora, Deezer, LastFM, etc.

Introduce Client::connecting state to better track status of a player's readiness to connect to a new stream.
Introduce Song:retryData, used opaquely to Song by StreamingController, to track progress of retry attempts.
Introduce StreamingController::isRetrying(), mainly so that the display code can put up the right text.

Slimproto:
If a DSCO with a reason code (error disconnect) is received while playing and when the player (decoder) buffer is empty then this is probably because a stream attempt (strm-s) that was sent in response to an STMd (that is, either streaming the next track or retrying the current one) has failed with a connection failure. A quirk in the firmware (both ip3k and SqueezePlay) means that, when the output buffer drains, the expected STMu will not be sent in this case. So, if the player is playing when the DSCO(error) is reported, and the decoder buffer is empty, then tell the controller that the player has stopped, synthesising an STMu.

StreamingController:
It was already the case that _RetryOrNext(), called for a ReadyToStream event, would attempt to retry a stream that it could identify (had been identified) as radio. This now sets up Song::retryData with initial values so that subsequent events may retry again if the initial reconnect fails.
The retry logic is encapsulated in _willRetry().
If the Song::open() called from _Stream() fails (only likely to happen for proxy streaming) then _willRetry() is called and, only if it returns false, is the usual failure path followed. Similarly for the two handlers called for a StreamingFailed event (_StopNextIfMore(), _SyncStopNext()), which is what one expects for direct streaming.
If the current streaming Song has retryData, then _willRetry() will set up a timed retry using the usual NextTrackReady event callback. This means that any user intervention that changes the playlist state will cancel the retries because it will invalidate the NextTrackReady callback when it occurs. Successfully starting to play also resets (undef) Song::retryData().


Bug: 3161
git-svn-id: file:///home/awy/mirror/slim/7.6/trunk/server@31371 62299810-d8cb-41fd-93b7-d32162e5a4a4